create ngff sub-package and refactor display util file
