+++ iohub/ngff/display.py
